feat(iroh-net): Add PkarrNodeDiscovery to iroh-net (#2628)
## Description

Deprecate iroh-pkarr-node-discovery and add the full dht based node
discovery to iroh itself.

## Breaking Changes

None

## Notes & open questions

PkarrNodeDiscovery does both resolution and publishing in one thing,
which is different to the two relay based versions.
Should we keep it this way? I think yes, given that we are instantiating
a full dht client and we don't want 2 of them...

Should we keep in the relay options, or remove them and make this one a
pure DHT one?

I copied the example from iroh-pkarr-node-discovery as well.

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] ~~Tests if relevant.~~
- [x] All breaking changes documented.