fix: Fix the task leak with the lazy in-mem rpc client while still keeping it lazy (#31)
* Make blobs more cheaply cloneable by by giving it an Inner

* Remove the lazy part.

The lazy handler kept a reference to Blobs alive. This caused both the
task and the blobs to never be dropped. To solve this you can just split
the inner part in 2 parts, one that has the handle and one that has the
logic. But that is not nice. I think it is best for the mem rpc handler
to exist completely separately, especially given that rpc is a non-default
feature.

* spawn_rpc should make it sufficiently clear that this is a thing you need to put away somewhere.

Or maybe spawn_client?

* back to the lazy client

* add comment about the purpose of the handler