feat: Enable RPC by default (#73)
## Description

Almost the entire friendly API of the crate is only available if the
`rpc` feature flag is set. So this enables rpc by default to make the
crate easier to use for beginners.

I also document feature flags in the module docs (should this instead be
in README.md?) so people that need fewer deps know what to do.

## Breaking Changes

None

## Notes & open questions

Note: there is also a bugfix to the new tags API. The rename fn in the
API was not actually using the shiny new atomic rename but just the
multi-step rename. If we decide not to enable rpc by default I will move
that change into a separate PR.

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.