refactor(net)!: Remove `Joined` event, use `NeighborUp` (#49)
## Description

This removes the `GossipEvent::Joined` and replaces it by emitting `NeighborUp`.
The `Joined` event was emitted as the first event when joining a topic, either once the first neighbor showed up, or, in case the topic was already joined, immediately with the list of current neighbors.

Semantically, there is no difference to the `NeighborUp`  event. I think it makes code using this code simpler and easier to understand if we remove the `Joined` event and just emit `NeighborUp` events as needed.

## Breaking Changes

* `iroh_gossip::net::GossipEvent::Joined` is removed. Instead, upon joining a topic or subscribing to a joined topic, `GossipEvent::NeighborUp` is emitted for each active neighbor. When using `GossipReceiver::joined` or `GossipTopic::joined`, you don't need to change anything, these functions work as they did previously.

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR. -->

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text), if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.