fix: getting deleted blobs should return error (#95)
## Description

Right now there is a bug in the store: When you delete a blob by
deleting its tag, run gc, and then do `store.get_bytes(hash)`, it will
not return an error but instead an emtpy byte slice (`b""`). This, of
course, is very wrong: we should never return data that doesn't match
the requested hash.

This PR fixes the logic to not create empty blobs on-demand when running
get or export operations.

Also adds a test that ensures that, for deleted blobs, errors are
returned for `get_bytes`, `export`, `export_bao`, `export_ranges` for
both the `fs` and `mem` store.

Additionally, I noticed that `Blobs::delete` and `delete_with_opts` is
not actually usable as a public API because there's internal protections
of hashes that make it not work as intended (it is only used from gc,
where protections are cleared before). Therefore I made the function
non-public to not confuse users.

## Breaking Changes

* `iroh_blobs::api::blobs::Blobs::delete` and `delete_with_opts` are no
longer public. These functions did not actually delete blobs, because
blobs would usually be protected. The function is now only used
internally during garbage collection (where protections are cleared
beforehand in a non-public way). Public users can only use garbage
collection for deletion of blobs.

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.

---------

Co-authored-by: Ruediger Klaehn <rklaehn@protonmail.com>