docs: Add module docs (#114)
## Description

Adds module docs and includes the README.md into the docs so the example
in README.md is being tested.

## Breaking Changes

None, I guess. The additional deref is not breaking.

## Notes & open questions

There is an unrelated change, adding back lifetimes, the lack of which
that make my local clippy unhappy (I only removed the lifetimes because
of clippy in the first place).

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.