feat: Read and seek (#119)
## Description

Add an api for reading and seeking a blob using tokio::io::AsyncRead and
tokio::io::AsyncSeek

## Breaking Changes

Changes api::blobs::ExportRangesProgress::stream to be non-async - it
does not have to be!

## Notes & open questions

Note: the entire machinery of seek and read is horribly undocumented
regarding the non happy case. So I went with the following approach: if
you ever drop a future of either seek or read before completing, the
reader becomes poisioned and will never work again, you have to grab a
new one, which is cheap.

This bypasses the whole question about what to do if somebody calls x
without having completed y, and in normal usage you should never notice
this.

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.