refactor: Remove slot mutex and simplify per blob state (#104)
## Description

Before, the state we kept per hash was a bit convoluted. The entry point
was a Slot, which contained a tokio mutex to cover the async loading of
an entry state from the metadata db. Then inside that, there was the
actual entry state, wrapped in an option to cover the case that we don't
have anything about the hash.

This was working, but it was an arc in an arc in an arc, and also led to
bugs in the case of trying to export a blob that doesn't exist.

Now all these states are flattened into a single enum, and we can easily
define what should happen when we e.g. do an export of an entry that
does not exist - return an appropriate io error.

## Breaking Changes

None

## Notes & open questions

Note: you could remove the Initial and Loading state by using a
tokio::sync::OnceCell. But that is a true sync primitive, and we want to
use an AtomicRefCell, and also using a OnceLock would come with its own
sync primitive (a semaphore) just for init, and then we have our own one
for the non-load state changes. I don't think it is that bad...

Note: a lot of changes are to add the wait_idle fn in the rpc protocol,
which isn't really related to the changes in the PR but just a way to
get rid of some flaky tests.

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.