feat: Use reflink_or_copy for export as well (#137)
## Description

Use reflink_or_copy when exporing, not just when importing.

## Breaking Changes

None

## Notes & open questions

Note: we lose progress, but for modern file systems it will be instant,
so 🤷 . If you are a poor person having to live with Fat32, you can
always use TryReference.

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.