refactor: Replace connection pool (#138)
## Description

Replaces the somewhat hackish connection pool with the one from
https://github.com/n0-computer/iroh-experiments/pull/36 that was battle
tested more.

## Breaking Changes

None

## Notes & open questions

Q: Expose the conn pool here?

Note: There is a nice list of possible extensions, but I think it is
probably best to first get the basic version in.

Extensions would be `async fn ban(node_id: NodeId, duration:
Option<Duration>)`, bans the node for a time, or for as long as the conn
pool lives if duration is set to None.

A way to observe pool stats so users know when to schedule new downloads
without having to try.

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.