feat: impl From for api::Store (#169)
## Description

Currently, when functions want a `iroh_blobs::api::Store`, and you have
a `MemStore` or `FsStore`, you have to do `fs_store.as_ref().clone()` to
get from the `FsStore` to the `api::Store`. I stumbled a couple of times
and had to look that up, it's not very intuitive and hard to find. So to
make this easier, this adds `From` impls for `api::Store` for the
various stores.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.