feat: upgrade database to redb v3 compatible format (#174)
## Description

The recently released redb v3 has incompatible changes to the database
format. The recommended upgrade procedure is to upgrade the database
with `Database::upgrade` on redb v2.6+. The upgrade has to be done on
redb v2, not v3. See
https://github.com/cberner/redb/blob/master/CHANGELOG.md#removes-support-for-file-format-v2.

This updates redb to 2.6 and performs the upgrade when opening the
database. Calling upgrade on an already upgraded database is a no-op.

With this merged in the next release, we can then update redb to v3 in
the version after.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.