deps: Fix deps to actually work, not just compile, in wasm-browser (#201)
## Description

While this will currently compile to wasm, when running in the browser,
any reference to `std::time` will explode with unsupported platform
errors, this fixes that.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

Note: replaces #194 which is obsolete due to the change to n0-error

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.